refactor to remove boiler plate rename components and move from views to components directory
