+++ src/routes.js
@@ -17,7 +17,7 @@ import Payment from "views/Payment/Payment.js";
 //core components/views for Home layout
 import LandingPage from "views/LandingPage/LandingPage.js";
 import Home from "views/Home/Home.js";
-import Search from "views/Search/Search.js";
+import SearchResult from "views/SearchResult/SearchResult.js";
 import VideoUpload from "./views/VideoUpload/VideoUpload";
 import SignIn from "views/Auth/SignIn.js";
 import SignUp from "views/Auth/SignUp.js";
@@ -99,7 +99,7 @@ const routes = [
   {
     path: "/search/:query",
     name: "Search Results",
-    component: Search,
+    component: SearchResult,
     layout: "/home",
     exact: true,
   },
